Add assertions for commands that use type
Add, invalid and mark commands use an integer type for certain
functions.

These integers have a set range of values. Falling outside of these
ranges will cause issues.

Let's add assertions to check ensure they always fall within
their values.

These variables are not influenced directly by the user, so any invalid
input is indicative of a program bug, not user error.

So, using an assertion to handle it seems appropriate, as opposed to
using it to check something like command typos, where we want the user
to be able to correct their command instead of closing down the program.