AddCommand: Change variable name and extract addTask method FindCommand: Fix find list not showing
AddCommand has some code in its execute method that was shared across
all Task cases.

This causes code duplication.

Let's pull out the AddCommand code that was common across cases.

This allows the execute method to be shorter and more readable

Let's also fix a small bug that causes the Find command to not show
the result list properly