Add reminder function
The user has to manually enter today's date to find tasks due today.

This might be a hassle since checking tasks due that day may be a
commonly repeated action.

Let's add an automatic reminder that tells the user tasks due today on
startup.

This allows the user to be more aware of their deadlines without
further input from them.