Set scroll pane to stretch with resizing Set minimum width and height for resizing
Scroll pane does not resize with window.
Window can be shrunk as small as possible.

This makes resizing look awkward when scroll pane doesn't follow the
window.
Window shrinking too small makes the text impossible to see and the app
unusable.

Let's make the scroll pane stretch with the window.
Let's limit the minimum shrinking of the window.

This allows the scroll pane to look good when resizing.
This prevents the window size from causing visual issues.