Add TextColourPair utility class and changed message colouring
Dodo messages only change colour based on command type.
This is due to colour information being passed through a private String
in each command class object.

Even if there is an error message, the colour is still based on the
command type. This is not very flexible.
Implementation also leads to a lot of redundant coding for commands
that don't need a colour change.

Let's use the UI class to return a TextColourPair class instead to pass
both a text String and a colour String.

This allows the text colour and contents to be centralised into one
class, and also lets the changing of individual messages colour under
more flexible circumstances.